projects/inoahbrain/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
chromadb>=0.4.0
